Move form data manipulation to formdetails.readable
